Move percentile rank logic to user model
