+++ app/controllers/users_controller.rb
@@ -5,17 +5,6 @@ class UsersController < ApplicationController
     send_cache_disablement_headers
 
     @actionPages = ActionPage.order('id desc')
-    @user_action_counts = Rails.cache.fetch('user_action_counts', :expires_in => 24.hours) {
-      User.select( "users.id, count(ahoy_events.id) AS events_count" )
-        .joins( "LEFT OUTER JOIN ahoy_events ON ahoy_events.user_id = users.id" )
-        .where( "ahoy_events.name IS null OR ahoy_events.name = ?", "Action" )
-        .group( "users.id" )
-        .map { |u| u.events_count }
-    }
-
-    user_count = current_user.events.actions.count
-    @percentile = @user_action_counts.percentile_rank(user_count-1).round(0).ordinalize
-
     @actions_taken = current_user.events.actions.map {|i| i['properties']['actionPageId'].to_i}
   end
