+++ app/views/users/show.html.erb
@@ -65,7 +65,7 @@
       <div class="action-stats">
         <h3>Stats</h3>
         <div class="percentile">
-          <p>You are in the <span><%= get_percentile_ordinalized() %> </span>percentile of users.</p>
+          <p>You are in the <span><%= current_user.percentile_rank.ordinalize %> </span>percentile of users.</p>
         </div>
         <p>Actions you've taken so far:</p>
         <div class="action-grid">
